requirements.txt
streamlit
transformers
torch
accelerate
sentencepiece
bitsandbytes>=0.46.1
google-generativeai